+++ deye/src/locker/deye_empty_locker.py
@@ -1,7 +1,7 @@
 from deye_base_locker import DeyeBaseLocker
 
 class DeyeEmptyLocker(DeyeBaseLocker):
-  def acquire(self, timeout: int = 15) -> None:
+  def acquire(self, timeout: float = 15) -> None:
     pass
 
   def release(self) -> None:
